Fix route nesting
