+++ front/app/containers/Admin/projects/components/ProjectRow/ManageButton.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-import { adminProjectsProjectPath } from 'containers/Admin/projects/routes';
-
 import { FormattedMessage } from 'utils/cl-intl';
 
 import messages from '../messages';
@@ -18,7 +16,7 @@ const ManageButton = ({ isDisabled, publicationId }: Props) => {
       className={`
       e2e-admin-edit-publication intercom-admin-project-edit-button
     `}
-      linkTo={adminProjectsProjectPath(publicationId)}
+      linkTo={`/admin/projects/${publicationId}/phases/setup`}
       buttonStyle="secondary"
       icon="edit"
       type="button"
