+++ front/app/containers/Admin/projects/routes.tsx
@@ -8,15 +8,21 @@ import PageLoading from 'components/UI/PageLoading';
 
 import { AdminRoute } from '../routes';
 
-import AdminProjectIdeaPreviewIndex from './AdminProjectIdeaPreviewIndex';
-import IdeaFormBuilder from './project/inputForm/IdeaFormBuilder';
-import SurveyFormBuilder from './project/nativeSurvey/SurveyFormBuilder';
-import ProjectParticipation from './project/participation';
-import ProjectTraffic from './project/traffic';
-
+const AdminProjectIdeaPreviewIndex = lazy(
+  () => import('./AdminProjectIdeaPreviewIndex')
+);
+const IdeaFormBuilder = lazy(
+  () => import('./project/inputForm/IdeaFormBuilder')
+);
+const SurveyFormBuilder = lazy(
+  () => import('./project/nativeSurvey/SurveyFormBuilder')
+);
+const ProjectParticipation = lazy(() => import('./project/participation'));
+const ProjectTraffic = lazy(() => import('./project/traffic'));
+const AdminProjectsProjectIndex = lazy(() => import('./project'));
 const AdminProjectsAndFolders = lazy(() => import('.'));
 const AdminProjectsList = lazy(() => import('./all'));
-const AdminProjectsProjectIndex = lazy(() => import('./project'));
+const AdminProjectPhaseIndex = lazy(() => import('./project/phase'));
 const AdminProjectsProjectSettings = lazy(() => import('./project/settings'));
 const AdminProjectsProjectGeneral = lazy(() => import('./project/general'));
 const AdminPhaseNewAndEdit = lazy(() => import('./project/timeline/edit'));
@@ -58,8 +64,8 @@ export enum projectsRoutes {
   projects = 'projects',
   projectIdeaId = ':projectId/ideas/:ideaId',
   projectSettings = ':projectId/settings',
-  projectTraffic = ':projectId/traffic',
-  projectParticipation = ':projectId/participation',
+  projectTraffic = 'traffic',
+  projectParticipation = 'participation',
   projectSettingsDescription = 'description',
   projectSettingsEvents = 'events',
   projectSettingsEventsNew = 'events/new',
@@ -67,25 +73,26 @@ export enum projectsRoutes {
   projectSettingsTags = 'tags',
   projectSettingsAccessRights = 'access-rights',
   projectId = ':projectId',
-  projectPhasesSetup = 'phases/setup',
-  projectPhaseSetup = 'phases/:phaseId/setup',
-  projectNewPhase = 'phases/new',
-  projectPhase = 'phases/:phaseId',
-  projectPhaseSurveyResults = 'phases/:phaseId/survey-results',
-  projectPhasePolls = 'phases/:phaseId/polls',
-  projectPhaseAccessRights = 'phases/:phaseId/access-rights',
-  projectPhaseIdeas = 'phases/:phaseId/ideas',
-  projectPhaseIdeaForm = 'phases/:phaseId/ideaform',
-  projectPhaseVolunteering = 'phases/:phaseId/volunteering',
-  projectPhaseMap = 'phases/:phaseId/map',
-  projectPhaseNativeSurvey = 'phases/:phaseId/native-survey',
-  projectPhaseVolunteeringNewCause = 'phases/:phaseId/volunteering/causes/new',
-  projectPhaseIdeaFormEdit = 'phases/:phaseId/ideaform/edit',
-  projectPhaseNativeSurveyEdit = 'phases/:phaseId/native-survey/edit',
-  projectPhaseVolunteeringCause = 'phases/:phaseId/volunteering/causes/:causeId',
+  projectIdPhases = 'phases',
+  projectPhasesSetup = 'setup',
+  projectPhaseSetup = ':phaseId/setup',
+  projectNewPhase = 'new',
+  projectPhase = ':phaseId',
+  projectPhaseSurveyResults = ':phaseId/survey-results',
+  projectPhasePolls = ':phaseId/polls',
+  projectPhaseAccessRights = ':phaseId/access-rights',
+  projectPhaseIdeas = ':phaseId/ideas',
+  projectPhaseIdeaForm = ':phaseId/ideaform',
+  projectPhaseVolunteering = ':phaseId/volunteering',
+  projectPhaseMap = ':phaseId/map',
+  projectPhaseNativeSurvey = ':phaseId/native-survey',
+  projectPhaseVolunteeringNewCause = ':phaseId/volunteering/causes/new',
+  projectPhaseIdeaFormEdit = ':phaseId/ideaform/edit',
+  projectPhaseNativeSurveyEdit = ':phaseId/native-survey/edit',
+  projectPhaseVolunteeringCause = ':phaseId/volunteering/causes/:causeId',
+  projectPhaseOfflineInputs = ':phaseId/offline-inputs',
+  projectPhaseReport = ':phaseId/report',
   projectAnalysis = 'analysis/:analysisId',
-  projectPhaseOfflineInputs = 'phases/:phaseId/offline-inputs',
-  projectPhaseReport = 'phases/:phaseId/report',
 }
 
 export type projectsRouteTypes =
@@ -97,6 +104,7 @@ export type projectsRouteTypes =
   | AdminRoute<`${projectsRoutes.projects}/${string}/${projectsRoutes.projectSettingsEvents}/${string}`>
   | AdminRoute<`${projectsRoutes.projects}/${string}/${projectsRoutes.projectSettingsTags}`>
   | AdminRoute<`${projectsRoutes.projects}/${string}`>
+  | AdminRoute<`${projectsRoutes.projects}/${string}/phases`>
   | AdminRoute<`${projectsRoutes.projects}/${string}/${projectsRoutes.projectTraffic}`>
   | AdminRoute<`${projectsRoutes.projects}/${string}/${projectsRoutes.projectParticipation}`>
   | AdminRoute<`${projectsRoutes.projects}/${string}/${projectsRoutes.projectPhasesSetup}`>
@@ -142,22 +150,6 @@ const createAdminProjectsRoutes = () => {
         element: (
           <PageLoading>
             <AdminProjectIdeaPreviewIndex />
-          </PageLoading>
-        ),
-      },
-      {
-        path: projectsRoutes.projectTraffic,
-        element: (
-          <PageLoading>
-            <ProjectTraffic />
-          </PageLoading>
-        ),
-      },
-      {
-        path: projectsRoutes.projectParticipation,
-        element: (
-          <PageLoading>
-            <ProjectParticipation />
           </PageLoading>
         ),
       },
@@ -233,165 +225,192 @@ const createAdminProjectsRoutes = () => {
         // all routes under /admin/projects/:projectId
         children: [
           {
-            path: '',
-            element: <Navigate to="phases/setup" replace />,
+            path: projectsRoutes.projectTraffic,
+            element: (
+              <PageLoading>
+                <ProjectTraffic />
+              </PageLoading>
+            ),
           },
+          {
+            path: projectsRoutes.projectParticipation,
+            element: (
+              <PageLoading>
+                <ProjectParticipation />
+              </PageLoading>
+            ),
+          },
+          {
+            path: projectsRoutes.projectIdPhases,
+            element: (
+              <PageLoading>
+                <AdminProjectPhaseIndex />
+              </PageLoading>
+            ),
+            // all routes under /admin/projects/:projectId/phases
+            children: [
+              {
+                path: '',
+                element: <Navigate to="phases/setup" replace />,
+              },
 
-          {
-            path: projectsRoutes.projectPhasesSetup,
-            element: (
-              <PageLoading>
-                <AdminPhaseNewAndEdit />
-              </PageLoading>
-            ),
+              {
+                path: projectsRoutes.projectPhasesSetup,
+                element: (
+                  <PageLoading>
+                    <AdminPhaseNewAndEdit />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseSetup,
+                element: (
+                  <PageLoading>
+                    {/* We use the key here to make sure that the component is treated as a different instance
+                    to differentiate between the new and edit phase. This distinction is especially important
+                    when the component is already visible and the route changes to the same component.
+                    For example, from phase setup to creating a new phase.
+                    */}
+                    <AdminPhaseNewAndEdit key="setup" />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectNewPhase,
+                element: (
+                  <PageLoading>
+                    {/* We use the key here to make sure that the component is treated as a different instance
+                    to differentiate between the new and edit phase. This distinction is especially important
+                    when the component is already visible and the route changes to the same component.
+                    For example, from phase setup to creating a new phase.
+                    */}
+                    <AdminPhaseNewAndEdit key="new" />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhase,
+                element: (
+                  <PageLoading>
+                    <AdminPhaseNewAndEdit />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseSurveyResults,
+                element: (
+                  <PageLoading>
+                    <AdminProjectSurveyResults />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhasePolls,
+                element: (
+                  <PageLoading>
+                    <AdminProjectPoll />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseAccessRights,
+                element: (
+                  <PageLoading>
+                    <AdminProjectPermissions />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseIdeas,
+                element: (
+                  <PageLoading>
+                    <AdminProjectIdeas />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseIdeaForm,
+                element: (
+                  <PageLoading>
+                    <AdminProjectIdeaForm />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseVolunteering,
+                element: (
+                  <PageLoading>
+                    <AdminProjectVolunteering />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseMap,
+                element: (
+                  <PageLoading>
+                    <AdminCustomMapConfigComponent />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseVolunteeringNewCause,
+                element: (
+                  <PageLoading>
+                    <AdminProjectVolunteeringNew />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseNativeSurvey,
+                element: <AdminProjectsSurvey />,
+              },
+              {
+                path: projectsRoutes.projectPhaseIdeaFormEdit,
+                element: <IdeaFormBuilder />,
+              },
+              {
+                path: projectsRoutes.projectPhaseNativeSurveyEdit,
+                element: <SurveyFormBuilder />,
+              },
+              {
+                path: projectsRoutes.projectPhaseVolunteeringCause,
+                element: (
+                  <PageLoading>
+                    <AdminProjectVolunteeringEdit />
+                  </PageLoading>
+                ),
+              },
+              // {
+              //   path: 'allowed-input-topics',
+              //   element: <AdminAllowedTopicsComponent />,
+              // },
+              {
+                path: projectsRoutes.projectAnalysis,
+                element: (
+                  <PageLoading>
+                    <AdminProjectAnalysis />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseOfflineInputs,
+                element: (
+                  <PageLoading>
+                    <OfflineInputImporter />
+                  </PageLoading>
+                ),
+              },
+              {
+                path: projectsRoutes.projectPhaseReport,
+                element: (
+                  <PageLoading>
+                    <ReportTab />
+                  </PageLoading>
+                ),
+              },
+              ...moduleConfiguration.routes['admin.projects.project'],
+            ],
           },
-          {
-            path: projectsRoutes.projectPhaseSetup,
-            element: (
-              <PageLoading>
-                {/* We use the key here to make sure that the component is treated as a different instance
-                to differentiate between the new and edit phase. This distinction is especially important
-                when the component is already visible and the route changes to the same component.
-                For example, from phase setup to creating a new phase.
-                */}
-                <AdminPhaseNewAndEdit key="setup" />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectNewPhase,
-            element: (
-              <PageLoading>
-                {/* We use the key here to make sure that the component is treated as a different instance
-                to differentiate between the new and edit phase. This distinction is especially important
-                when the component is already visible and the route changes to the same component.
-                For example, from phase setup to creating a new phase.
-                */}
-                <AdminPhaseNewAndEdit key="new" />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhase,
-            element: (
-              <PageLoading>
-                <AdminPhaseNewAndEdit />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseSurveyResults,
-            element: (
-              <PageLoading>
-                <AdminProjectSurveyResults />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhasePolls,
-            element: (
-              <PageLoading>
-                <AdminProjectPoll />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseAccessRights,
-            element: (
-              <PageLoading>
-                <AdminProjectPermissions />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseIdeas,
-            element: (
-              <PageLoading>
-                <AdminProjectIdeas />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseIdeaForm,
-            element: (
-              <PageLoading>
-                <AdminProjectIdeaForm />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseVolunteering,
-            element: (
-              <PageLoading>
-                <AdminProjectVolunteering />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseMap,
-            element: (
-              <PageLoading>
-                <AdminCustomMapConfigComponent />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseVolunteeringNewCause,
-            element: (
-              <PageLoading>
-                <AdminProjectVolunteeringNew />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseNativeSurvey,
-            element: <AdminProjectsSurvey />,
-          },
-          {
-            path: projectsRoutes.projectPhaseIdeaFormEdit,
-            element: <IdeaFormBuilder />,
-          },
-          {
-            path: projectsRoutes.projectPhaseNativeSurveyEdit,
-            element: <SurveyFormBuilder />,
-          },
-          {
-            path: projectsRoutes.projectPhaseVolunteeringCause,
-            element: (
-              <PageLoading>
-                <AdminProjectVolunteeringEdit />
-              </PageLoading>
-            ),
-          },
-          // {
-          //   path: 'allowed-input-topics',
-          //   element: <AdminAllowedTopicsComponent />,
-          // },
-          {
-            path: projectsRoutes.projectAnalysis,
-            element: (
-              <PageLoading>
-                <AdminProjectAnalysis />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseOfflineInputs,
-            element: (
-              <PageLoading>
-                <OfflineInputImporter />
-              </PageLoading>
-            ),
-          },
-          {
-            path: projectsRoutes.projectPhaseReport,
-            element: (
-              <PageLoading>
-                <ReportTab />
-              </PageLoading>
-            ),
-          },
-          ...moduleConfiguration.routes['admin.projects.project'],
         ],
       },
     ],
